toml setup for memilio plot
